+++ Assets/Editor/Scripts/Metadata.cs
@@ -8,11 +8,11 @@ namespace Thisaislan.PersistenceEasyToDeleteInEditor.Editor.Constants
         internal const string PedeSettingFolderPath = "Assets/Settings";
         internal const string PedeSettingFileName = "PedeSettings.asset";
         
-        internal const string PedeDataFolderPath = "Assets/PEDE";
+        internal const string PedeDataFolderPath = "Assets/Pede";
         internal const string PedeDataFileName = "PedeData.asset";
         
         internal const string PedeDataScriptableObjectFileName = "PedeData";
-        internal const string PedeDataScriptableObjectMenuName = "PEDE/Data";
+        internal const string PedeDataScriptableObjectMenuName = "Pede/Data";
         internal const int PedeDataScriptableObjectOrder = 1;
         
         internal const int PedeDataValueMinLines = 4;